requirements.txt
Flask==2.3.3
Flask-Cors==3.0.10
mongoengine==0.27.0
pymongo==4.6.0
python-dotenv==1.0.0
requests==2.31.0
pytz==2023.3
flask-jwt-extended==4.7.1
flask-bcrypt==1.0.1
pytest==7.4.3
